Fix Relationship Database, v1.

+++ src/main/java/com/example/starwars/infrastructure/repository/impl/FindApplicablePriceImpl.java
@@ -35,7 +35,7 @@ public class FindApplicablePriceImpl implements FindApplicablePrice {
         OffsetDateTime offsetDateTime = localDateTime.atOffset(ZoneOffset.UTC);
         return priceRepository.findByProductIdAndBrandIdAndDate(
                 requestEntityEntity.getProductId(),
-                requestEntityEntity.getBrandId(),
+                (long) requestEntityEntity.getBrandId(),
                 offsetDateTime
         );
     }
